+++ poetry/console/commands/env/info.py
@@ -1,0 +1,60 @@
+import os
+import sys
+
+from ..command import Command
+
+
+class EnvInfoCommand(Command):
+    """
+    Display information about the current environment.
+
+    env:info
+        {--p|path : Only display the environment's path}
+    """
+
+    def handle(self):
+        from poetry.utils.env import Env
+
+        poetry = self.poetry
+        env = Env.get(cwd=poetry.file.parent)
+
+        if self.option("path"):
+            if not env.is_venv():
+                return 1
+
+            self.write(str(env.path))
+
+            return
+
+        self._display_complete_info(env)
+
+    def _display_complete_info(self, env):
+        env_python_version = ".".join(str(s) for s in env.version_info[:3])
+        self.output.title("Virtualenv")
+        listing = [
+            "<info>Python</info>:         <comment>{}</>".format(env_python_version),
+            "<info>Implementation</info>: <comment>{}</>".format(
+                env.python_implementation
+            ),
+            "<info>Path</info>:           <comment>{}</>".format(
+                env.path if env.is_venv() else "NA"
+            ),
+        ]
+        if env.is_venv():
+            listing.append(
+                "<info>Valid</info>:          <{tag}>{is_valid}</{tag}>".format(
+                    tag="comment" if env.is_sane() else "error", is_valid=env.is_sane()
+                )
+            )
+        self.output.listing(listing)
+
+        self.line("")
+
+        self.output.title("System")
+        self.output.listing(
+            [
+                "<info>Platform</info>: <comment>{}</>".format(env.platform),
+                "<info>OS</info>:       <comment>{}</>".format(env.os),
+                "<info>Python</info>:   <comment>{}</>".format(env.base),
+            ]
+        )